refactor: extract pure version helpers into dfyb.version with tests
